+++ components/detail/realtime_search_detail.py
